src/main/webapp/app/entities/feuillage/route/feuillage-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IFeuillage, Feuillage } from '../feuillage.model';
import { FeuillageService } from '../service/feuillage.service';

import { FeuillageRoutingResolveService } from './feuillage-routing-resolve.service';

describe('Feuillage routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: FeuillageRoutingResolveService;
  let service: FeuillageService;
  let resultFeuillage: IFeuillage | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(FeuillageRoutingResolveService);
    service = TestBed.inject(FeuillageService);
    resultFeuillage = undefined;
  });

  describe('resolve', () => {
    it('should return IFeuillage returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultFeuillage = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultFeuillage).toEqual({ id: 123 });
    });

    it('should return new IFeuillage if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultFeuillage = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultFeuillage).toEqual(new Feuillage());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Feuillage })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultFeuillage = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultFeuillage).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
